+++ oci/layer/generate_linux_test.go
@@ -82,3 +82,103 @@ func TestTranslateOverlayWhiteouts_Char00(t *testing.T) {
 		})
 	})
 }
+
+func TestTranslateOverlayWhiteouts_XattrOpaque(t *testing.T) {
+	dir := t.TempDir()
+
+	testNeedsTrustedOverlayXattrs(t)
+
+	err := os.Mkdir(filepath.Join(dir, "wodir"), 0755)
+	require.NoError(t, err)
+	err = unix.Lsetxattr(filepath.Join(dir, "wodir"), "trusted.overlay.opaque", []byte("y"), 0)
+	require.NoError(t, err, "lsetxattr trusted.overlay.opaque")
+	err = os.WriteFile(filepath.Join(dir, "reg"), []byte("dummy file"), 0644)
+	require.NoError(t, err)
+
+	packOptions := RepackOptions{TranslateOverlayWhiteouts: true}
+
+	t.Run("GenerateLayer", func(t *testing.T) {
+		// something reasonable
+		mtreeKeywords := []mtree.Keyword{
+			"size",
+			"type",
+			"uid",
+			"gid",
+			"mode",
+		}
+		deltas, err := mtree.Check(dir, nil, mtreeKeywords, fseval.Default)
+		assert.NoError(t, err, "mtree check")
+
+		reader, err := GenerateLayer(dir, deltas, &packOptions)
+		assert.NoError(t, err, "generate layer")
+		defer reader.Close()
+
+		checkLayerEntries(t, reader, []tarDentry{
+			{path: ".", ftype: tar.TypeDir},
+			{path: "reg", ftype: tar.TypeReg, contents: "dummy file"},
+			{path: "wodir/", ftype: tar.TypeDir},
+			{path: "wodir/" + whOpaque, ftype: tar.TypeReg},
+		})
+	})
+
+	t.Run("GenerateInsertLayer", func(t *testing.T) {
+		reader := GenerateInsertLayer(dir, "/", false, &packOptions)
+		defer reader.Close()
+
+		checkLayerEntries(t, reader, []tarDentry{
+			{path: "/", ftype: tar.TypeDir},
+			{path: "reg", ftype: tar.TypeReg, contents: "dummy file"},
+			{path: "wodir/", ftype: tar.TypeDir},
+			{path: "wodir/" + whOpaque, ftype: tar.TypeReg},
+		})
+	})
+}
+
+func TestTranslateOverlayWhiteouts_XattrWhiteout(t *testing.T) {
+	dir := t.TempDir()
+
+	testNeedsTrustedOverlayXattrs(t)
+
+	err := os.WriteFile(filepath.Join(dir, "woreg"), []byte{}, 0755)
+	require.NoError(t, err)
+	err = unix.Lsetxattr(filepath.Join(dir, "woreg"), "trusted.overlay.whiteout", []byte("foobar"), 0)
+	require.NoError(t, err, "lsetxattr trusted.overlay.whiteout")
+	err = os.WriteFile(filepath.Join(dir, "reg"), []byte("dummy file"), 0644)
+	require.NoError(t, err)
+
+	packOptions := RepackOptions{TranslateOverlayWhiteouts: true}
+
+	t.Run("GenerateLayer", func(t *testing.T) {
+		// something reasonable
+		mtreeKeywords := []mtree.Keyword{
+			"size",
+			"type",
+			"uid",
+			"gid",
+			"mode",
+		}
+		deltas, err := mtree.Check(dir, nil, mtreeKeywords, fseval.Default)
+		assert.NoError(t, err, "mtree check")
+
+		reader, err := GenerateLayer(dir, deltas, &packOptions)
+		assert.NoError(t, err, "generate layer")
+		defer reader.Close()
+
+		checkLayerEntries(t, reader, []tarDentry{
+			{path: ".", ftype: tar.TypeDir},
+			{path: "reg", ftype: tar.TypeReg, contents: "dummy file"},
+			{path: whPrefix + "woreg", ftype: tar.TypeReg},
+		})
+	})
+
+	t.Run("GenerateInsertLayer", func(t *testing.T) {
+		reader := GenerateInsertLayer(dir, "/", false, &packOptions)
+		defer reader.Close()
+
+		checkLayerEntries(t, reader, []tarDentry{
+			{path: "/", ftype: tar.TypeDir},
+			{path: "reg", ftype: tar.TypeReg, contents: "dummy file"},
+			{path: whPrefix + "woreg", ftype: tar.TypeReg},
+		})
+	})
+}